backend/requirements.txt
ai21==3.3.0
ai21-tokenizer==0.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiolimiter==1.2.1
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.13.0
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncmy==0.2.10
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.35
bcrypt==4.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
boto3==1.38.33
botocore==1.38.33
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.8
cloudpickle==3.1.1
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==41.0.7
cycler==0.12.1
dask==2025.5.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
deepmerge==2.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
diskcache==5.6.3
distributed==2025.5.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.1.0
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastavro==1.11.1
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gpt4all==2.8.2
greenlet==3.2.3
grpcio
grpcio-status
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_ai==2.31.5
jupyter_ai_magics==2.31.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-aws==0.2.24
langchain-cohere==0.4.4
langchain-community==0.3.24
langchain-core==0.3.64
langchain-google-genai==2.1.5
langchain-mistralai==0.2.10
langchain-nvidia-ai-endpoints==0.3.10
langchain-ollama==0.3.3
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
locket==1.0.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.22.0
openai==1.85.0
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
posthog==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf
psutil==7.0.0
psycopg2-binary==2.9.10
pulsar-client==3.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyMySQL==1.1.0
pyodbc==4.0.39
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.0
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qianfan==0.4.12.3
qtconsole==5.6.1
QtPy==2.4.3
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
seaborn==0.13.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
structlog==23.2.0
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
together==1.5.11
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.15.4
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.24.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
zstandard==0.23.0
